requirements.txt
#https://devcenter.heroku.com/articles/python-support
Cython==0.29.15
dash==1.9.1
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
dash-table-experiments==0.6.0
docutils==0.15.2
ephem==3.7.7.1
Flask==1.0.2
Flask-Compress==1.4.0
Frozen-Flask==0.15
future==0.18.2
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
jedi==0.15.1
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
lazy-object-proxy==1.4.2
LunarCalendar==0.0.9
pandas==0.25.3
Pillow==6.2.0
plotly==4.5.4
pylint==2.4.2
pyparsing==2.4.6
pystan==2.19.1.1
python-dateutil==2.8.0
pytz==2019.3
setuptools-git==1.2
Werkzeug==1.0.0
six==1.12.0
